Move session_stream files to session files
